feat: Add TodoStatusParam for App Intents, enabling initial status selection in AddTodoIntent and centralizing the status enum.
