+++ TodoMate/AppIntent/Intent/AddTodoIntent.swift
@@ -22,10 +22,14 @@ struct AddTodoIntent: AppIntent {
   @Parameter(title: "Date", default: .now)
   var date: Date
 
+  @Parameter(title: "Status", default: .todo)
+  var status: TodoStatusParam
+
   static var parameterSummary: some ParameterSummary {
     Summary("Add \(\.$content)") {
       \.$date
       \.$detail
+      \.$status
     }
   }
 
@@ -36,10 +40,12 @@ struct AddTodoIntent: AppIntent {
     }
 
     // 로컬 Todo는 owner를 ""로 설정 (AppIntent는 Firebase 미사용)
+    let sanitizedContent = content.replacingOccurrences(of: "\n", with: " ")
 
     let todo = Todo(
       owner: "",
-      content: content,
+      content: sanitizedContent,
+      status: status.asDomain,
       detail: detail ?? "",
       in: date,
     )
