+++ TodoMate/AppIntent/Intent/UpdateTodoIntent.swift
@@ -22,28 +22,6 @@ struct UpdateTodoIntent: AppIntent {
   @Parameter(title: "New Content")
   var newContent: String?
 
-  // AppEnum for TodoStatus
-  enum TodoStatusParam: String, AppEnum {
-    case todo = "시작 전"
-    case inProgress = "진행 중"
-    case complete = "완료"
-
-    static var typeDisplayRepresentation: TypeDisplayRepresentation = "Todo Status"
-    static var caseDisplayRepresentations: [TodoStatusParam: DisplayRepresentation] = [
-      .todo: "To Do",
-      .inProgress: "In Progress",
-      .complete: "Complete",
-    ]
-
-    var asDomain: TodoStatus {
-      switch self {
-      case .todo: .todo
-      case .inProgress: .inProgress
-      case .complete: .complete
-      }
-    }
-  }
-
   @MainActor
   func perform() async throws -> some IntentResult & ReturnsValue<TodoEntity> {
     guard let container = CoreDIContainer.shared else {
